Внесены изменения по замечаниям ревьюера

+++ src/main/java/ru/yandex/practicum/filmorate/service/MappingManager.java
@@ -6,9 +6,6 @@ import ru.yandex.practicum.filmorate.model.User;
 
 public interface MappingManager {
     boolean validateFilm(Film film);
-    //имеет ли смысл делать обобщенные методы по типу <T> boolean validate(T object)
-    //а далее в реализации через instanceof производить каст объекта? Или лучше сразу написать несколько реализаций
-    //схожих методов, как мы это делали с тасками-сабтасками-эпиками
 
     boolean validateUpdateFilm(Film film);
 
@@ -16,7 +13,7 @@ public interface MappingManager {
 
     boolean validateUpdateUser(User user);
 
-    int nextFilmId();
+    Long nextFilmId();
 
-    int nextUserId();
+    Long nextUserId();
 }
